backend/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.0
pydantic>=2.7.0
openai>=1.42.0
python-dotenv>=1.0.1
python-docx>=1.1.0
xhtml2pdf>=0.2.11
jinja2>=3.1.3
python-multipart>=0.0.9
